requirements.txt
flask
requests
openai
python-dotenv
flask-cors
run it on a disabled web security chrome

